requirements.txt
# Core Dependencies
docling>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
supabase>=2.0.0

# Production Dependencies
python-multipart>=0.0.6
slowapi>=0.1.9
tenacity>=8.2.3
prometheus-client>=0.19.0

# Authentication & Security
pyjwt>=2.8.0
bcrypt>=4.1.2
python-jose[cryptography]>=3.3.0

# Payment Processing
stripe>=7.0.0

# Async Support
httpx>=0.26.0
aiofiles>=23.2.1
